move to namespace-type packaging
