+++ examples/core/pipe_example.yml
@@ -9,14 +9,18 @@ site:
   max_threads: 2
 stages:
 - classname: FlowEngine
+  module_name: rail.creation.engines.flowEngine
   name: flow_engine_test
   nprocess: 1
 - classname: LSSTErrorModel
+  module_name: rail.creation.degradation.lsst_error_model
   name: lsst_error_model_test
   nprocess: 1
 - classname: ColumnMapper
+  module_name: rail.core.utilStages
   name: col_remapper_test
   nprocess: 1
 - classname: TableConverter
+  module_name: rail.core.utilStages
   name: table_conv_test
   nprocess: 1
