--- rail/__init__.py
@@ -1,8 +1,0 @@
-"""RAIL, the Redshift Assesement Infrastructre Layers"""
-
-from .version import __version__
-from . import core
-from . import creation
-from . import evaluation
-from . import estimation
-from .main import *
