+++ rail/core/__init__.py
@@ -1,7 +1,15 @@
 """Core code for RAIL"""
 
-from .stage import *
-from .data import *
-from .utilStages import *
-from .utilPhotometry import *
-from .common_params import *
+__all__ = ["__version__"]
+
+import os
+from importlib.metadata import PackageNotFoundError, version
+
+__version__: str
+"""The application version string (PEP 440 / SemVer compatible)."""
+
+try:
+    __version__ = version(__name__)
+except PackageNotFoundError:
+    # package is not installed
+    __version__ = "0.0.0"
