+++ rail/core/utils.py
@@ -1,0 +1,6 @@
+""" Utility functions """
+
+import os
+import rail.core
+
+RAILDIR = os.path.abspath(os.path.join(os.path.dirname(rail.core.__file__), '..', '..'))
