--- rail/creation/__init__.py
@@ -1,3 +1,0 @@
-"""Modules to create synthetic data"""
-from .degradation import *
-from .engines import *
