--- rail/creation/engines/__init__.py
@@ -1,4 +1,0 @@
-"""Engines that create synthetic samples of photometric data"""
-
-from .engine import *
-from .flowEngine import *
