--- rail/estimation/__init__.py
@@ -1,2 +1,0 @@
-"""Redshift estimation"""
-from .algos import *
