+++ rail/estimation/algos/flexzboost.py
@@ -144,7 +144,7 @@ class Inform_FZBoost(CatInformer):
 
         if self.config.hdf5_groupname:
             training_data = self.get_data('input')[self.config.hdf5_groupname]
-        else:  #pragma:  no cover
+        else:  #pragma: no cover
             training_data = self.get_data('input')
         speczs = training_data['redshift']
         print("stacking some data...")
