+++ rail/estimation/algos/sklearn_nn.py
@@ -94,7 +94,7 @@ class Inform_SimpleNN(CatInformer):
         import sklearn.neural_network as sknn
         if self.config.hdf5_groupname:
             training_data = self.get_data('input')[self.config.hdf5_groupname]
-        else:  #pragma:  no cover
+        else:  #pragma: no cover
             training_data = self.get_data('input')
         speczs = training_data['redshift']
         print("stacking some data...")
