+++ rail/estimation/algos/trainZ.py
@@ -39,7 +39,7 @@ class Inform_trainZ(CatInformer):
     def run(self):
         if self.config.hdf5_groupname:
             training_data = self.get_data('input')[self.config.hdf5_groupname]
-        else:  #pragma:  no cover
+        else:  #pragma: no cover
             training_data = self.get_data('input')
         zbins = np.linspace(self.config.zmin, self.config.zmax, self.config.nzbins+1)
         speczs = np.sort(training_data['redshift'])
