--- rail/evaluation/__init__.py
@@ -1,3 +1,0 @@
-"""Tools to evaluate performance of Estimators """
-from .evaluator import Evaluator
-from .metrics import *
